Code review

+++ backend/kendo-tournament-core/src/test/java/com/softwaremagico/kt/core/tests/SimpleChampionshipTest.java
@@ -203,7 +203,7 @@ public class SimpleChampionshipTest extends AbstractTestNGSpringContextTests {
             }
         }
 
-        Assert.assertEquals((int) TEAMS, teamProvider.count(tournament));
+        Assert.assertEquals(TEAMS, teamProvider.count(tournament));
     }
 
     @Test(dependsOnMethods = {"addTeams"})
